Portaudio integration 0.2.0
* add portaudio (wip)

* fix play example

* add todos

* fix examples

* start cpal_backend (WIP)

* more work in cpal integration

* update feature names

* slight redesign for dynamic processor use in same engine

* remove stopped trait function again

* include audio backends in neoaudio

* rename crate error to Error

* fix documentation improve error

* remove all unwraps

* remove unwraps and fix example, version 0.2.0

* ignore manual test

* clippy fix

+++ crates/neo-audio/src/device_config.rs
@@ -5,8 +5,8 @@ pub struct DeviceConfig {
     pub api: String,
     pub output_device: Device,
     pub input_device: Device,
-    pub num_output_ch: u32,
-    pub num_input_ch: u32,
+    pub num_output_ch: u16,
+    pub num_input_ch: u16,
     pub sample_rate: u32,
     pub num_frames: u32,
 }
